+++ src/reducer.js
@@ -6,8 +6,8 @@ import { fromJS } from 'immutable'
 import { normalize } from 'normalizr'
 
 import {
-  ACTION_DELETE_TYPE_NAME,
   ACTION_EMPTY_TYPE_NAME,
+  ACTION_REMOVE_TYPE_NAME,
   ACTION_TYPE_PREFIX,
   ACTIONS_REDUCER_NAME,
   ENTITIES_REDUCER_NAME
@@ -33,6 +33,7 @@ const makeActionsReducer = defaultActionsState => {
     }
 
     const {
+      args,
       status,
       time,
       actionId,
@@ -48,8 +49,11 @@ const makeActionsReducer = defaultActionsState => {
 
     const entityKey = actionSchema.key
 
-    // action.clear
-    if (action.type === ACTION_EMPTY_TYPE_NAME) {
+    // action.clear || action.remove
+    if (
+      action.type === ACTION_EMPTY_TYPE_NAME ||
+      action.type === ACTION_REMOVE_TYPE_NAME
+    ) {
       return state.set(
         actionId,
         fromJS(Object.assign({ entityKey }, actionDefaultData, { time }))
@@ -68,13 +72,8 @@ const makeActionsReducer = defaultActionsState => {
       actionState = state.get(actionId)
     }
 
-    // delete entity id from actions
-    if (action.type === ACTION_DELETE_TYPE_NAME) {
-      console.log('action delete', actionState)
-      return state
-    }
-
     actionState = actionState.merge({
+      args,
       status,
       time,
       hasError,
@@ -142,12 +141,6 @@ const makeEntitiesReducer = defaultEntitiesState => {
     const newEntitiesItems = normalizedPayloadSource
       ? normalizedPayloadSource.entities
       : {}
-
-    // delete entity id from actions
-    if (action.type === ACTION_DELETE_TYPE_NAME) {
-      console.log('reducer delete')
-      return state
-    }
 
     // merge entity item data
     for (let entityName in newEntitiesItems) {